+++ exinst/default.nix
@@ -1,17 +1,19 @@
-{ mkDerivation, aeson, base, bytes, constraints, generic-random
-, profunctors, QuickCheck, singletons, stdenv, tasty, tasty-hunit
-, tasty-quickcheck
+{ mkDerivation, aeson, base, bytes, constraints, deepseq
+, generic-random, hashable, profunctors, QuickCheck, singletons
+, stdenv, tasty, tasty-hunit, tasty-quickcheck
 }:
 mkDerivation {
   pname = "exinst";
   version = "0.2";
   src = ./.;
   libraryHaskellDepends = [
-    aeson base bytes constraints profunctors QuickCheck singletons
+    aeson base bytes constraints deepseq hashable profunctors
+    QuickCheck singletons
   ];
   testHaskellDepends = [
-    aeson base bytes constraints generic-random profunctors QuickCheck
-    singletons tasty tasty-hunit tasty-quickcheck
+    aeson base bytes constraints deepseq generic-random hashable
+    profunctors QuickCheck singletons tasty tasty-hunit
+    tasty-quickcheck
   ];
   homepage = "https://github.com/k0001/exinst";
   description = "Derive instances for your existential types";